refactor: #76 더미데이터 정리
