+++ src/lib/dummy/wiki.ts
@@ -263,3 +263,4 @@ export function getDummyWikiBySlug(slug: string): DummyWiki {
     content,
   };
 }
+
